install from fork

+++ run.py
@@ -8,22 +8,18 @@ import time
 
 if __name__ == '__main__':
     path = Path.home()/'.kaggle'
-    os.mkdir(path)    
     
-    with open(path/'kaggle.json', 'w') as f: f.write(os.environ['key'])
+    if not path.exists():
+        os.mkdir(path)    
+        with open(path/'kaggle.json', 'w') as f: f.write(os.environ['key'])
 
     libs = L(open('libraries.txt').read().split('\n')).filter(lambda x: x != '')
     lib_path = Path('./kaggle_datasets')
     
-#    create_libs_datasets(libs,lib_path=lib_path,username='isaacflath',clear_after=False)
-
-    f = bind(create_libs_datasets,lib_path=lib_path,username='isaacflath',clear_after=True)
-    def tryf(lib):
+    for lib in libs:
         try:
-            f(lib)
+            create_libs_datasets(lib,lib_path=lib_path,username='isaacflath',clear_after=False)
         except:
-            print(f"API limit hit - sleep and retry....")
+            print('API limit hit - 30 second hold')
             time.sleep(30)
-            f(lib)
-    
-    parallel(tryf,libs)
+            create_libs_datasets(lib,lib_path=lib_path,username='isaacflath',clear_after=False)
